use uint64_t instead of size_t in UDPConnectionStats, because there's no overload of Pair() constructor for size_t

+++ src/udpapi.h
@@ -22,14 +22,14 @@ enum UDPConnectionType {
 };
 
 // fUltimatelyTrusted means you trust them (ie whitelist) and ALL OF THEIR SUBSEQUENT WHITELISTED PEERS
-void OpenUDPConnectionTo(const CService& remote_addr, uint64_t local_magic, uint64_t remote_magic, bool fUltimatelyTrusted, UDPConnectionType connection_type = UDP_CONNECTION_TYPE_NORMAL, size_t group = 0);
-void OpenPersistentUDPConnectionTo(const CService& remote_addr, uint64_t local_magic, uint64_t remote_magic, bool fUltimatelyTrusted, UDPConnectionType connection_type = UDP_CONNECTION_TYPE_NORMAL, size_t group = 0);
+void OpenUDPConnectionTo(const CService& remote_addr, uint64_t local_magic, uint64_t remote_magic, bool fUltimatelyTrusted, UDPConnectionType connection_type = UDP_CONNECTION_TYPE_NORMAL, uint64_t group = 0);
+void OpenPersistentUDPConnectionTo(const CService& remote_addr, uint64_t local_magic, uint64_t remote_magic, bool fUltimatelyTrusted, UDPConnectionType connection_type = UDP_CONNECTION_TYPE_NORMAL, uint64_t group = 0);
 
 void CloseUDPConnectionTo(const CService& remote_addr);
 
 struct UDPConnectionStats {
     CService remote_addr;
-    size_t group;
+    uint64_t group;
     bool fUltimatelyTrusted;
     int64_t lastRecvTime;
     std::vector<double> last_pings;
